% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluviumplot.R
\name{plot_alluvium}
\alias{plot_alluvium}
\title{Alluvial Plot Function}
\usage{
plot_alluvium(df, x_axis, grouping, weights, plot_title, xlab, ylab,
  axis_breaks, has_lines = FALSE, line_breaks = NULL, add_values = FALSE)
}
\arguments{
\item{df}{A dataframe containing each variable as a column; must be in molted form (see reshape package for details)}

\item{x_axis}{the column in df containing the x axis information}

\item{grouping}{the column in df containing the group identity information}

\item{weights}{the column containing the y axis information}

\item{plot_title}{str; title of your plot}

\item{xlab}{str; x-axis label}

\item{ylab}{str; y-axis label}

\item{axis_breaks}{vector of how the x axis should be partitioned}

\item{has_lines}{if TRUE, includes vertical line breaks separating along the axis_breaks}

\item{line_breaks}{a vector that contains the x-values for where the lines should be manually placed. Only functional if has_lines = TRUE}

\item{add_values}{if TRUE, includes values of the weights corresponding to the alluvium strata}
}
\description{
This function allows you to plot an alluvial/Sankey diagram
}
\examples{
plot_alluvium(df = melt_df, x_axis = melt_df$year, grouping = melt_df$factors, 
weights = melt_df$value, plot_title = "Risk Factors for Stroke in Blacks", 
xlab = "Year", ylab = "Cumulative Proportion", axis_breaks = yearaxes)
}
\keyword{Sankey}
\keyword{alluvium}
\keyword{plot}
